host/gui/requirements.txt
pyserial
PyQt6>=6.5
